+++ Proline-ORM/src/main/java/fr/proline/core/orm/pdi/BioSequence.java
@@ -47,11 +47,11 @@ public class BioSequence implements Serializable {
 
     private String crc64;
 
-    private Integer length;
+    private int length;
 
-    private double mass;
+    private int mass;
 
-    private float pi;
+    private Float pi;
 
     private String sequence;
 
@@ -89,27 +89,27 @@ public class BioSequence implements Serializable {
 	this.crc64 = crc64;
     }
 
-    public Integer getLength() {
+    public int getLength() {
 	return this.length;
     }
 
-    public void setLength(Integer length) {
+    public void setLength(int length) {
 	this.length = length;
     }
 
-    public double getMass() {
+    public int getMass() {
 	return this.mass;
     }
 
-    public void setMass(double mass) {
+    public void setMass(int mass) {
 	this.mass = mass;
     }
 
-    public float getPi() {
+    public Float getPi() {
 	return this.pi;
     }
 
-    public void setPi(float pi) {
+    public void setPi(Float pi) {
 	this.pi = pi;
     }
